+++ packages/remix-dev/__tests__/readConfig-test.ts
@@ -1,17 +1,23 @@
 import path from "path";
 
 import type { RemixConfig } from "../config";
-import { readConfig } from "../config";
+import { serverBuildTargetWarning, readConfig } from "../config";
 
 const remixRoot = path.resolve(__dirname, "./fixtures/stack");
 
 describe("readConfig", () => {
   let config: RemixConfig;
+  let warnStub;
   beforeEach(async () => {
+    let consoleWarn = console.warn;
+    warnStub = jest.fn();
+    console.warn = warnStub;
     config = await readConfig(remixRoot);
+    console.warn = consoleWarn;
   });
 
   it("generates a config", async () => {
+    expect(warnStub).toHaveBeenCalledWith(serverBuildTargetWarning);
     expect(config).toMatchInlineSnapshot(
       {
         rootDirectory: expect.any(String),
@@ -21,10 +27,14 @@ describe("readConfig", () => {
         serverBuildPath: expect.any(String),
         assetsBuildDirectory: expect.any(String),
         relativeAssetsBuildDirectory: expect.any(String),
+        entryClientFilePath: expect.any(String),
+        entryServerFilePath: expect.any(String),
         tsconfigPath: expect.any(String),
         future: {
           unstable_cssModules: expect.any(Boolean),
           unstable_cssSideEffectImports: expect.any(Boolean),
+          unstable_postcss: expect.any(Boolean),
+          unstable_tailwind: expect.any(Boolean),
           unstable_vanillaExtract: expect.any(Boolean),
           v2_errorBoundary: expect.any(Boolean),
           v2_meta: expect.any(Boolean),
@@ -39,11 +49,15 @@ describe("readConfig", () => {
         "devServerBroadcastDelay": 0,
         "devServerPort": Any<Number>,
         "entryClientFile": "entry.client.tsx",
+        "entryClientFilePath": Any<String>,
         "entryServerFile": "entry.server.tsx",
+        "entryServerFilePath": Any<String>,
         "future": Object {
           "unstable_cssModules": Any<Boolean>,
           "unstable_cssSideEffectImports": Any<Boolean>,
           "unstable_dev": false,
+          "unstable_postcss": Any<Boolean>,
+          "unstable_tailwind": Any<Boolean>,
           "unstable_vanillaExtract": Any<Boolean>,
           "v2_errorBoundary": Any<Boolean>,
           "v2_meta": Any<Boolean>,
@@ -61,10 +75,16 @@ describe("readConfig", () => {
           },
         },
         "serverBuildPath": Any<String>,
-        "serverBuildTarget": undefined,
+        "serverBuildTarget": "node-cjs",
         "serverBuildTargetEntryModule": "export * from \\"@remix-run/dev/server-build\\";",
+        "serverConditions": undefined,
         "serverDependenciesToBundle": Array [],
         "serverEntryPoint": undefined,
+        "serverMainFields": Array [
+          "main",
+          "module",
+        ],
+        "serverMinify": false,
         "serverMode": "production",
         "serverModuleFormat": "cjs",
         "serverPlatform": "node",