+++ packages/remix-dev/compiler/compileBrowser.ts
@@ -28,10 +28,13 @@ import {
 } from "./plugins/cssBundleEntryModulePlugin";
 import { writeFileSafe } from "./utils/fs";
 import invariant from "../invariant";
+import { hmrPlugin } from "./plugins/hmrPlugin";
 
 export type BrowserCompiler = {
   // produce ./public/build/
-  compile: (manifestChannel: WriteChannel<AssetsManifest>) => Promise<void>;
+  compile: (
+    manifestChannel: WriteChannel<AssetsManifest>
+  ) => Promise<esbuild.Metafile>;
   dispose: () => void;
 };
 
@@ -76,10 +79,11 @@ const isCssBundlingEnabled = (config: RemixConfig) =>
 const createEsbuildConfig = (
   build: "app" | "css",
   config: RemixConfig,
-  options: CompileOptions
+  options: CompileOptions,
+  onLoader: (filename: string, code: string) => void
 ): esbuild.BuildOptions | esbuild.BuildIncremental => {
   let isCssBuild = build === "css";
-  let entryPoints: esbuild.BuildOptions["entryPoints"];
+  let entryPoints: Record<string, string>;
 
   if (isCssBuild) {
     entryPoints = {
@@ -87,7 +91,7 @@ const createEsbuildConfig = (
     };
   } else {
     entryPoints = {
-      "entry.client": path.resolve(config.appDirectory, config.entryClientFile),
+      "entry.client": config.entryClientFilePath,
     };
 
     for (let id of Object.keys(config.routes)) {
@@ -99,7 +103,6 @@ const createEsbuildConfig = (
   }
 
   let { mode } = options;
-  let { rootDirectory } = config;
   let outputCss = isCssBuild;
 
   let plugins: esbuild.Plugin[] = [
@@ -108,21 +111,41 @@ const createEsbuildConfig = (
       ? cssBundleEntryModulePlugin(config)
       : null,
     config.future.unstable_cssModules
-      ? cssModulesPlugin({ mode, rootDirectory, outputCss })
+      ? cssModulesPlugin({ config, mode, outputCss })
       : null,
     config.future.unstable_vanillaExtract
       ? vanillaExtractPlugin({ config, mode, outputCss })
       : null,
     config.future.unstable_cssSideEffectImports
-      ? cssSideEffectImportsPlugin({ rootDirectory })
+      ? cssSideEffectImportsPlugin({ config, options })
       : null,
-    cssFilePlugin({ mode, rootDirectory }),
+    cssFilePlugin({ config, options }),
     urlImportsPlugin(),
     mdxPlugin(config),
-    browserRouteModulesPlugin(config, /\?browser$/),
+    browserRouteModulesPlugin(config, /\?browser$/, onLoader, mode),
     emptyModulesPlugin(config, /\.server(\.[jt]sx?)?$/),
     NodeModulesPolyfillPlugin(),
   ].filter(isNotNull);
+
+  if (mode === "development" && config.future.unstable_dev) {
+    // TODO prebundle deps instead of chunking just these ones
+    let isolateChunks = [
+      "react",
+      "react/jsx-dev-runtime",
+      "react/jsx-runtime",
+      "react-dom",
+      "react-dom/client",
+      "react-refresh/runtime",
+      "@remix-run/react",
+      "remix:hmr",
+    ];
+    entryPoints = {
+      ...entryPoints,
+      ...Object.fromEntries(isolateChunks.map((imprt) => [imprt, imprt])),
+    };
+
+    plugins.push(hmrPlugin({ remixConfig: config }));
+  }
 
   return {
     entryPoints,
@@ -165,6 +188,9 @@ const createEsbuildConfig = (
     jsx: "automatic",
     jsxDev: options.mode !== "production",
     plugins,
+    supported: {
+      "import-meta": true,
+    },
   };
 };
 
@@ -175,11 +201,18 @@ export const createBrowserCompiler = (
   let appCompiler: esbuild.BuildIncremental;
   let cssCompiler: esbuild.BuildIncremental;
 
+  let hmrRoutes: Record<string, { loaderHash: string }> = {};
+  let onLoader = (filename: string, code: string) => {
+    let key = path.relative(remixConfig.rootDirectory, filename);
+    hmrRoutes[key] = { loaderHash: code };
+  };
+
   let compile = async (manifestChannel: WriteChannel<AssetsManifest>) => {
+    hmrRoutes = {};
     let appBuildTask = async () => {
       appCompiler = await (!appCompiler
         ? esbuild.build({
-            ...createEsbuildConfig("app", remixConfig, options),
+            ...createEsbuildConfig("app", remixConfig, options, onLoader),
             metafile: true,
             incremental: true,
           })
@@ -189,6 +222,7 @@ export const createBrowserCompiler = (
         appCompiler.metafile,
         "Expected app compiler metafile to be defined. This is likely a bug in Remix. Please open an issue at https://github.com/remix-run/remix/issues/new"
       );
+      return appCompiler.metafile;
     };
 
     let cssBuildTask = async () => {
@@ -200,7 +234,7 @@ export const createBrowserCompiler = (
       //  so we need to assert that it's an incremental build
       cssCompiler = (await (!cssCompiler
         ? esbuild.build({
-            ...createEsbuildConfig("css", remixConfig, options),
+            ...createEsbuildConfig("css", remixConfig, options, onLoader),
             metafile: true,
             incremental: true,
             write: false,
@@ -235,11 +269,6 @@ export const createBrowserCompiler = (
 
       let cssBundlePath = cssBundleFile.path;
 
-      // Get esbuild's existing CSS source map so we can pass it to PostCSS
-      let cssBundleSourceMap = outputFiles.find((outputFile) =>
-        isCssBundleFile(outputFile, ".css.map")
-      )?.text;
-
       let { css, map } = await postcss([
         // We need to discard duplicate rules since "composes"
         // in CSS Modules can result in duplicate styles
@@ -247,8 +276,10 @@ export const createBrowserCompiler = (
       ]).process(cssBundleFile.text, {
         from: cssBundlePath,
         to: cssBundlePath,
-        map: {
-          prev: cssBundleSourceMap,
+        map: options.sourcemap && {
+          prev: outputFiles.find((outputFile) =>
+            isCssBundleFile(outputFile, ".css.map")
+          )?.text,
           inline: false,
           annotation: false,
           sourcesContent: true,
@@ -274,15 +305,39 @@ export const createBrowserCompiler = (
       return cssBundlePath;
     };
 
-    let [cssBundlePath] = await Promise.all([cssBuildTask(), appBuildTask()]);
+    let [cssBundlePath, metafile] = await Promise.all([
+      cssBuildTask(),
+      appBuildTask(),
+    ]);
+
+    let hmr: AssetsManifest["hmr"] | undefined = undefined;
+    if (options.mode === "development" && remixConfig.future.unstable_dev) {
+      let hmrRuntimeOutput = Object.entries(metafile.outputs).find(
+        ([_, output]) => output.inputs["hmr-runtime:remix:hmr"]
+      )?.[0];
+      invariant(hmrRuntimeOutput, "Expected to find HMR runtime in outputs");
+      let hmrRuntime =
+        remixConfig.publicPath +
+        path.relative(
+          remixConfig.assetsBuildDirectory,
+          path.resolve(hmrRuntimeOutput)
+        );
+      hmr = {
+        runtime: hmrRuntime,
+        routes: hmrRoutes,
+        timestamp: Date.now(),
+      };
+    }
 
     let manifest = await createAssetsManifest({
       config: remixConfig,
       metafile: appCompiler.metafile!,
       cssBundlePath,
+      hmr,
     });
-    manifestChannel.write(manifest);
     await writeAssetsManifest(remixConfig, manifest);
+    manifestChannel.write(manifest);
+    return metafile;
   };
 
   return {