+++ packages/create-remix/CHANGELOG.md
@@ -1,5 +1,19 @@
 # `create-remix`
 
+## 1.14.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/dev@1.14.0`
+
+## 1.13.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/dev@1.13.0`
+
 ## 1.12.0
 
 ### Patch Changes